uim-vi/vim: m17n-im-config can configure IM in m17n
Thanks to *lewtds* from the team Bogo. The tool is very good, and
it can be used to control the variables in m17n. However, the tool
is out-of-date (last updated in 2007; m17n is update in Sep 2012).
It would be nice if some framework like ibus, scim, uim can control
the behavior of m17n from their side. It is too noisy to use two
different GUI tools to just one thing