rss: per-year support, add item link
As a work-around, the item link is the link to the commit on Github

We now support per-year new pages: news/2013, news/2014. However,
for the first year 2013 we are still using the page news/index.html,
because the ruby library :ruby-git does not support "--follow" option.

We need to update the "post-receive" githook get new feature to work