License change (CC BY-NC-ND -> CC BY-SA)
NC/ND is not a free license. That prevents the project from being
improved, distributed (commerically or not), and the project will
be rejected by most free/open source project hosting (not Github).

Special thanks to #tuxfamily for pointing out the issue, for their
great support and discussion.

Please note that we are supporting multiple licenses, and the default
license is CC BY-SA. To specify what license to be used, pls. use
update the field :license in your article. Use :on to disable