news: Goodbye :phant6m
We are sorry to announce the death of :phant6m.

Only one week after being launched, :phant6m was shut down
due to the Nari storm. However, due to some unknown reason,
the server was not lauched again.

The team decided to stop using :phant6m in the future. We
need a robust and stable system for our works.

Special thank #DSP team for their great support.