style.css: reduce menu font size. fix mobile padding
The font size for #cssmenu is now 12px (former is 13px), which is
smaller than the normal text size. This will help to keep menu
on a single line.

Actually, it is padding (left, right) settings for mobile users
that breaks #cssmenu into lines. By default, padding left is >0,
so for mobile users I add padding right >0 (10px) to give a balance
view. However, this will not give much space for the #cssmenu.
Now mobile users will not see any padding on left or right.