rss.rb: Use :news/index if file of the year missing
On the :master branch, we have news page for every year (2013,
2014). This is not true on :core branch whose contents are out
of date. We will check if the file of the year does exist, if
not we use the default index page. Please note that, the page
may exist but it does not belong to the git repository -- this
case still causes a Nanoc error