d.uim-vi.testing.2013: update #viber section.
Viber was built with Qt 5.3.1. It can not work with the latest version
of uim-vi (built with Qt 5.4.x). We need to copy an old plugin from
the known working version of uim-vi (ec1fc40).

This means it is hard to fix bug (if any) in the old version.
We may need a way to build uim-vi with different Qt.